Applied the rule of structure. Changed a couple of errs
This is a great attempt. It's great that you did this. Just for further work, Markdown works best (in fact any sort of markup, like HTML) works best when you use it only to define the structure of the document.  Structure is different from appearance. So using ###### to get really small bold text, isn't the right way, ###### defines a 6th level heading and should be used to denote that. The actual appearance comes later. Not a major thing at this stage, but it will be important to bear in mind.